+++ backend/src/main/java/com/api/backend/global/config/CorsConfig.java
@@ -1,0 +1,29 @@
+package com.api.backend.global.config;
+
+import java.util.Arrays;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
+
+@Configuration
+public class CorsConfig {
+
+  @Value("${frontend.host}")
+  String host;
+  @Value("${frontend.port}")
+  String port;
+
+  @Bean
+  public CorsFilter corsFilter() {
+    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+    CorsConfiguration config = new CorsConfiguration();
+    config.setAllowCredentials(true);
+    config.addAllowedOriginPattern("http://" + host + ":" + port);
+    config.addAllowedHeader("*");
+    config.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE"));
+    source.registerCorsConfiguration("/**", config);
+    return new CorsFilter(source);
+  }}
