+++ backend/src/main/java/com/api/backend/global/config/WebConfig.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
 import org.springframework.data.web.SortHandlerMethodArgumentResolver;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -33,20 +34,5 @@ public class WebConfig implements WebMvcConfigurer {
     pageableHandlerMethodArgumentResolver.setMaxPageSize(500);
     pageableHandlerMethodArgumentResolver.setFallbackPageable(PageRequest.of(0, 10));
     argumentResolvers.add(pageableHandlerMethodArgumentResolver);
-  }
-
-  @Bean
-  public CorsFilter corsFilter() {
-    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-    CorsConfiguration config = new CorsConfiguration();
-    config.setAllowCredentials(true);
-    config.addAllowedOriginPattern("http://" + host + ":" + port);
-    config.addAllowedHeader("*");
-    config.addAllowedMethod("GET");
-    config.addAllowedMethod("POST");
-    config.addAllowedMethod("PUT");
-    config.addAllowedMethod("DELETE");
-    source.registerCorsConfiguration("/**", config);
-    return new CorsFilter(source);
   }
 }
